src/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.1
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
dnspython==2.4.2
exceptiongroup==1.2.0
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
langchain-experimental==0.0.49
langchain-openai==0.0.2.post1
langsmith==0.0.77
loguru==0.7.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.3
openai==1.6.1
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pinecone-client==2.2.4
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.3
tenacity==8.2.3
tiktoken==0.5.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
yarl==1.9.4
